Added 2 web scraping projects
Camping cancellations ended up being a bit more complex than what I was looking for in the first project doing web scraping so it just reads campsite names. Coming soon is a fully functioning program that takes the movies coming soon on IMDB and outputs a calendar format of the release dates.